[feat] 댓글 api 직접 연동 주석

+++ src/api/CommentAPI.js
@@ -24,9 +24,9 @@ export const addCommentRequest = async (recruit_id, commentInfo) => {
 }
 
 // 댓글 삭제
-export const deleteCommentRequest = async (commentId) => {
+export const deleteCommentRequest = async (commentId, userId) => {
     try {
-        const response = await client.delete(`/comments/${commentId}`);
+        const response = await client.delete(`/comments/${commentId}`, userId);
         return response.data;
     } catch (error) {
         console.log("댓글 삭제에 실패했습니다.", error);
@@ -47,12 +47,12 @@ export const addReplyRequest = async (parnetId, replyInfo) => {
 }
 
 // 대댓글 삭제
-export const deleteReplyRequest = async (childCommentId) => {
+export const deleteReplyRequest = async (childCommentId, userId) => {
     try {
-        const response = await client.delete(`/child-comments/${childCommentId}`);
+        const response = await client.delete(`/child-comments/${childCommentId}`, userId);
         return response.data;
     } catch (error) {
-
+        console.log('대댓글 삭제에 실패했습니다.', error);
     }
 }
